Skip VimEnter autocmd when `g:barbaric_default` is not set (closes #26)
from @bluelightning32: upgrading from vim 9.1.1914-1.fc43 to 9.1.1952-1
resulted in a regression whereby the vim "splash screen" in the terminal
is suspended (i.e., not visible) until after the first keypress.

The cause was determined to be the VimEnter autocmd
defined at the end of `plugin/barbaric.vim`.
Further analysis revealed that this autocmd is largely unnecessary;
it only serves a purpose when both

1. `g:barbaric_default` is user-defined, and
2. vim is launched with an input method other than the above.

In this commit, we disable the VimEnter autocmd
except in the cases where it might possibly be relevant.

https://github.com/rlue/vim-barbaric/issues/26